pruning base to common tx, splite special function to devliver
